[features] Split regions.hpp in base and derived files #883
