+++ src/openMVG_Samples/multiview_robust_essential/robust_essential.cpp
@@ -8,7 +8,6 @@
 
 #include "openMVG/cameras/Camera_Pinhole.hpp"
 #include "openMVG/features/feature.hpp"
-#include "openMVG/features/regions_factory.hpp"
 #include "openMVG/features/svg_features.hpp"
 #include "openMVG/geometry/pose3.hpp"
 #include "openMVG/image/image_io.hpp"
