+++ src/openMVG_Samples/multiview_robust_homography_guided/robust_homography_guided.cpp
@@ -8,7 +8,6 @@
 
 #include "openMVG/features/feature.hpp"
 #include "openMVG/features/svg_features.hpp"
-#include "openMVG/features/regions_factory.hpp"
 #include "openMVG/image/image_io.hpp"
 #include "openMVG/image/image_concat.hpp"
 #include "openMVG/numeric/eigen_alias_definition.hpp"
